pyproject.toml
[tool.poetry]
name = "pych9329"
version = "0.1.5"
description = "Python module to control the ch9329"
authors = ["wevsty <ty@wevs.org>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/wevsty/pych9329"

[tool.poetry.dependencies]
python = ">3.10"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
pytest-datadir = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
